+++ lib/PuppeteerSharp/TimeoutSettings.cs
@@ -1,15 +1,24 @@
 namespace PuppeteerSharp
 {
-    internal class TimeoutSettings
+    /// <summary>
+    /// Timeout settings.
+    /// </summary>
+    public class TimeoutSettings
     {
         private int? _defaultNavigationTimeout;
 
+        /// <summary>
+        /// Navigation Timeout.
+        /// </summary>
         public int NavigationTimeout
         {
             get => _defaultNavigationTimeout ?? Timeout;
             set => _defaultNavigationTimeout = value;
         }
 
+        /// <summary>
+        /// Default timeout.
+        /// </summary>
         public int Timeout { get; set; } = Puppeteer.DefaultTimeout;
     }
 }
